ParseSyntaxFiles.hs: Allow regeneration of individual syntax files.
The purpose is to not have to regenerate (and more significantly,
recompile) all syntax files when modifying only one or a few.

With the change ParseSyntaxFiles accepts any number of arguments.
Those ending with .xml are assumed to be files while the rest are
assumed to be directories containing .xml files.

The syntaxes included in Syntax.hs are based on the contents of
destDir rather than the arguments, this allows regenerating a subset
of syntaxes but still having all syntaxes in Syntax.hs.