New, compressed format for classes.
Two-letter class names are now used for default styles, and detailed
lexical information is omitted unless the new OptDetailed option is
specified. Also, the "Normal" style is no longer used, and spans
are not used for "normal" markup.  The result is significant
compression of the highlighted source (in one test, 191K -> 72K).

NOTE: you must update your highlighting CSS stylesheets to
use the new two-letter identifiers.  There are new versions of
the three standard CSS files in the css directory of the
source distribution.

A new -d/--detailed option has been provided in the Highlight
executable. This selects OptDetailed.