API changes: highlight (and highlightAs) now lazily returns a list of SourceLine.
Previously, the entire input would have to be parsed before any input
was returned.  Now, we parse line by line.

Performance is approximately the same.

Code is somewhat simpler.

If a parsing error occurs, the offending line is highlighted as
ErrorTok, and we continue with the next line without change of state.
(This may produce funny results, but at least the offending line
will be clearly indicated.)

If you call highlightAs with a language that isn't supported,
all lines will be highlighted as NormalTok.