Improved handling of include directives, with API changes.
- Context is now a (String, String) rather than a String.
  The first is the syntax name, the second the context
  name.
- ContextStack is now a list of Contexts.
- synStLanguage has been removed.
- pEndLine parameter has been removed from mkParseSourceLine.
- parseExpression now handles the endline.
- parseExpressionInternal has been removed in the
  individual parsers; we now use parseExpression.

When we encounter an include, we try running parseExpression
for the target syntax with the initial context.  If that
succeeds, it might push a new context with the new syntax as
first component.  In this case, the parser will next call
parseExpression for the new syntax.  This will continue until
we are back to the original context.