Put code blocks in a div.
This reverts the change in a9c79de477397db38d790a01295c0689dcc10550
wrapping all blocks in a pre element.  (That was bad because pres
are only supposed to contain flow-level content.)

Instead we now wrap in a div.

The default CSS now includes "x-overflow: auto" for the div,
so that a scroll bar will appear on small screens.

See jgm/pandoc#1903 and jgm/highlighting-kate#65 for background.